dominium-56 Command Gateway und Query Gateway zur Wiederverwendung vorbereiten Refactoring: Inventur verwendet Epics statt auf fremde Actions zu hören
