--- ui/frontend/src/main/src/app/store/actions.ts
@@ -1,4 +1,0 @@
-export enum Actions {
-    InventurBegonnen = 'InventurBegonnen',
-    InventarErfasst = 'InventarErfasst'
-}
